\name{trades}
\alias{calculate.avg.price}
\alias{get.trade.data}
\alias{show.last.trades}
\alias{trades}
\title{Calculate trades}
\arguments{
  \item{typeID}{typeID for the item to be manufactured}
}
\description{
  This function reads in the entries from the CSV trade
  file downloads and returns a data.table of all the unique
  entries
}
\examples{
\dontrun{

}
}
\keyword{trades}

